+++ integration/integration_test_utils.py
@@ -179,7 +179,7 @@ class TestContext:
                 **kwargs,
             )
         elif self.driver_name == "SQLDriver" and IsAggregationlessDriver:
-            sql_driver_class = storey.drivers.SQLDriver
+            sql_driver_class = storey.sql_driver.SQLDriver
             return sql_driver_class(db_path=SQL_DB, primary_key=primary_key)
         elif self.driver_name == "SQLDriver" and not IsAggregationlessDriver:
             pytest.skip("SQLDriver doesnt support aggregation")
